Revert "Created a Command-Line Interface"
